src/requirements.txt
Flask==2.0.1
flask-bootstrap==3.3.7
Pillow==8.3.1
pytesseract==0.3.8
